chore: Engine v2: add selection set walkers (without plan), needed for caching
